Enable sessionID access in socket client with no change in frontend
This means it is a secure authentication that can be used to give the socket client
exactly the permissions that were granted on the express server with no option for
anyone to manipulate without having access to the server code.

Also the require statement was replaced with a default ES6 import.